[integration-cli] handle errors when removing
Handles two errors when removing all containers.

The first is if the container doesn't exist, which happens
if the test kills a container which exits between
getContainers and deleteContainers. In this case we ignore it.

The second is if a container removal is already in progress,
in which case we sleep for a second, and then add the container
for removal again.

Signed-off-by: Christopher Jones <tophj@linux.vnet.ibm.com>